+++ src/ios/UtilsPlugin.h
@@ -1,0 +1,7 @@
+#import <Cordova/CDVPlugin.h>
+
+@interface UtilsPlugin : CDVPlugin
+
+- (void) isApnsProduction: (CDVInvokedUrlCommand*)command;
+
+@end